feat(frontend): portable validation UX — FieldError, FormField, validateDogForm (US-035)
