+++ frontend/src/types/dog.ts
@@ -1,0 +1,6 @@
+export interface DogFormValues {
+  name: string;
+  breed: string;
+  dateOfBirth: string;
+  sex: string;
+}
